+++ src/main/java/edu/ifal/virtable/domain/SistemaRPG.java
@@ -1,19 +1,19 @@
 package edu.ifal.virtable.domain;
 
 public class SistemaRPG {
-    private int id;
+    private Long id;
 
     private String nome;
 
     private String descricao;
 
-    public SistemaRPG(int id, String nome, String descricao) {
+    public SistemaRPG(Long id, String nome, String descricao) {
         this.id = id;
         this.nome = nome;
         this.descricao = descricao;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
